requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv_python==4.10.0.82
Pillow==10.3.0
scipy==1.13.1
skimage==0.0
Werkzeug==3.0.3
